requirements.txt
# Core Flask dependencies
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0

# Async processing
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Data validation and processing
pydantic==2.5.0
email-validator==2.1.1
requests==2.31.0
pillow==10.0.1

# Environment and configuration
python-dotenv==1.0.0

# Utilities
unicodedata2==15.1.0

# Production dependencies
gunicorn==21.2.0
gevent==23.9.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Security
cryptography==41.0.0
bcrypt==4.1.2

# Development dependencies (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
